# core-lending-provisioning-risk-interfaces/src/main/java/com/firefly/core/lending/provisioning/interfaces/dtos/provisioning/v1/ProvisioningCaseDTO.java
/*
 * Copyright 2025 Firefly Software Foundation
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */


package com.firefly.core.lending.provisioning.interfaces.dtos.provisioning.v1;

import com.firefly.core.lending.provisioning.interfaces.enums.provisioning.v1.ProvisioningStatusEnum;
import com.firefly.core.lending.provisioning.interfaces.enums.provisioning.v1.RiskGradeEnum;
import com.firefly.core.lending.provisioning.interfaces.enums.assessment.v1.StageCodeEnum;
import org.fireflyframework.utils.annotations.FilterableId;
import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.PositiveOrZero;
import jakarta.validation.constraints.Size;
import java.math.BigDecimal;
import java.time.LocalDateTime;
import java.util.UUID;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class ProvisioningCaseDTO {
    @JsonProperty(access = JsonProperty.Access.READ_ONLY)
    private UUID provisioningCaseId;

    @FilterableId
    @NotNull(message = "Loan servicing case ID is required")
    private UUID loanServicingCaseId;  // External reference to Servicing

    @NotNull(message = "Stage code is required")
    private StageCodeEnum stageCode;   // STAGE_1, STAGE_2, STAGE_3, POCI

    @PositiveOrZero(message = "ECL amount must be positive or zero")
    private BigDecimal eclAmount;      // Current ECL amount

    @NotNull(message = "Risk grade is required")
    private RiskGradeEnum riskGrade;   // Risk grade (AAA, BBB, etc.)

    private LocalDateTime lastCalculatedAt;

    @NotNull(message = "Provisioning status is required")
    private ProvisioningStatusEnum provisioningStatus; // ACTIVE, RELEASED, etc.

    @Size(max = 1000, message = "Remarks cannot exceed 1000 characters")
    private String remarks;

    @JsonProperty(access = JsonProperty.Access.READ_ONLY)
    private LocalDateTime createdAt;

    @JsonProperty(access = JsonProperty.Access.READ_ONLY)
    private LocalDateTime updatedAt;
}