# core-lending-provisioning-risk-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-lending-provisioning-risk
  description: Provisioning Risk Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/loan-origination
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
  /api/v1/provisioning-cases/{provisioningCaseId}:
    get:
      tags:
        - ProvisioningCase
      summary: Get a provisioning case by ID
      operationId: getById
      parameters:
        - name: provisioningCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProvisioningCaseDTO'
    put:
      tags:
        - ProvisioningCase
      summary: Update a provisioning case
      operationId: update
      parameters:
        - name: provisioningCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProvisioningCaseDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProvisioningCaseDTO'
    delete:
      tags:
        - ProvisioningCase
      summary: Delete a provisioning case
      operationId: delete
      parameters:
        - name: provisioningCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/provisioning-cases/{provisioningCaseId}/calculations/{provisioningCalculationId}:
    get:
      tags:
        - ProvisioningCalculation
      summary: Get a provisioning calculation by ID
      operationId: getById_3
      parameters:
        - name: provisioningCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: provisioningCalculationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProvisioningCalculationDTO'
    put:
      tags:
        - ProvisioningCalculation
      summary: Update a provisioning calculation
      operationId: update_3
      parameters:
        - name: provisioningCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: provisioningCalculationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProvisioningCalculationDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProvisioningCalculationDTO'
    delete:
      tags:
        - ProvisioningCalculation
      summary: Delete a provisioning calculation
      operationId: delete_3
      parameters:
        - name: provisioningCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: provisioningCalculationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/provisioning-cases:
    get:
      tags:
        - ProvisioningCase
      summary: List or search provisioning cases
      operationId: findAll
      parameters:
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.loanServicingCaseId
          in: query
          description: Exact filter for loan servicing case id
          required: false
          schema:
            type: string
        - name: filters.stageCode
          in: query
          description: Exact filter for stage code
          required: false
          schema:
            type: string
            enum:
              - STAGE_1
              - STAGE_2
              - STAGE_3
              - POCI
        - name: filters.eclAmount
          in: query
          description: Exact filter for ecl amount
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[eclAmount].from
          in: query
          description: Filter ecl amount from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[eclAmount].to
          in: query
          description: Filter ecl amount to value
          required: false
          schema:
            type: string
        - name: filters.riskGrade
          in: query
          description: Exact filter for risk grade
          required: false
          schema:
            type: string
            enum:
              - AAA
              - AA
              - A
              - BBB
              - BB
              - B
              - CCC
              - CC
              - C
              - D
        - name: filters.lastCalculatedAt
          in: query
          description: Exact filter for last calculated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[lastCalculatedAt].from
          in: query
          description: Filter last calculated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[lastCalculatedAt].to
          in: query
          description: Filter last calculated at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.provisioningStatus
          in: query
          description: Exact filter for provisioning status
          required: false
          schema:
            type: string
            enum:
              - ACTIVE
              - RELEASED
              - WRITTEN_OFF
              - RECOVERED
        - name: filters.remarks
          in: query
          description: Exact filter for remarks
          required: false
          schema:
            type: string
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseProvisioningCaseDTO'
    post:
      tags:
        - ProvisioningCase
      summary: Create a new provisioning case
      operationId: create
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProvisioningCaseDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProvisioningCaseDTO'
...
  /api/v1/provisioning-cases/{provisioningCaseId}/calculations:
    get:
      tags:
        - ProvisioningCalculation
      summary: List or search provisioning calculations for a case
      operationId: findAll_3
      parameters:
        - name: provisioningCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.provisioningCaseId
          in: query
          description: Exact filter for provisioning case id
          required: false
          schema:
            type: string
        - name: filters.riskAssessmentId
          in: query
          description: Exact filter for risk assessment id
          required: false
          schema:
            type: string
        - name: filters.finalEcl
          in: query
          description: Exact filter for final ecl
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[finalEcl].from
          in: query
          description: Filter final ecl from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[finalEcl].to
          in: query
          description: Filter final ecl to value
          required: false
          schema:
            type: string
        - name: filters.calcMethod
          in: query
          description: Exact filter for calc method
          required: false
          schema:
            type: string
            enum:
              - TWELVE_MONTH_ECL
              - LIFETIME_ECL
              - EXPERT_OVERRIDE
        - name: filters.calcTimestamp
          in: query
          description: Exact filter for calc timestamp
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[calcTimestamp].from
          in: query
          description: Filter calc timestamp from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[calcTimestamp].to
          in: query
          description: Filter calc timestamp to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.notes
          in: query
          description: Exact filter for notes
          required: false
          schema:
            type: string
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseProvisioningCalculationDTO'
    post:
      tags:
        - ProvisioningCalculation
      summary: Create a new provisioning calculation
      operationId: create_3
      parameters:
        - name: provisioningCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProvisioningCalculationDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProvisioningCalculationDTO'
...
components:
  schemas:
    ProvisioningCaseDTO:
      required:
        - loanServicingCaseId
        - provisioningStatus
        - riskGrade
        - stageCode
      type: object
      properties:
        provisioningCaseId:
          type: string
          format: uuid
          readOnly: true
        loanServicingCaseId:
          type: string
          format: uuid
        stageCode:
          type: string
          enum:
            - STAGE_1
            - STAGE_2
            - STAGE_3
            - POCI
        eclAmount:
          type: number
        riskGrade:
          type: string
          enum:
            - AAA
            - AA
            - A
            - BBB
            - BB
            - B
            - CCC
            - CC
            - C
            - D
        lastCalculatedAt:
          type: string
          format: date-time
        provisioningStatus:
          type: string
          enum:
            - ACTIVE
            - RELEASED
            - WRITTEN_OFF
            - RECOVERED
        remarks:
          maxLength: 1000
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
...
    ProvisioningCalculationDTO:
      required:
        - calcMethod
        - calcTimestamp
        - finalEcl
        - provisioningCaseId
        - riskAssessmentId
      type: object
      properties:
        provisioningCalculationId:
          type: string
          format: uuid
          readOnly: true
        provisioningCaseId:
          type: string
          format: uuid
        riskAssessmentId:
          type: string
          format: uuid
        finalEcl:
          type: number
        calcMethod:
          type: string
          enum:
            - TWELVE_MONTH_ECL
            - LIFETIME_ECL
            - EXPERT_OVERRIDE
        calcTimestamp:
          type: string
          format: date-time
        notes:
          maxLength: 1000
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
...
    PaginationResponseProvisioningCaseDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/ProvisioningCaseDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...
    PaginationResponseProvisioningCalculationDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/ProvisioningCalculationDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...